requirements.txt
streamlit
google-genai
groq
gtts
deep-translator
pillow
python-dotenv